requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.0
pydantic>=2.0.0
PyYAML>=6.0
